+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/creation/customerlist/CustomerListViewModel.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.orders.creation.customerlistnew
+package com.woocommerce.android.ui.orders.creation.customerlist
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -6,10 +6,10 @@ import androidx.lifecycle.SavedStateHandle
 import com.woocommerce.android.R
 import com.woocommerce.android.analytics.AnalyticsEvent
 import com.woocommerce.android.analytics.AnalyticsTrackerWrapper
-import com.woocommerce.android.ui.orders.creation.customerlistnew.CustomerListGetSupportedSearchModes.Companion.SEARCH_MODE_VALUE_ALL
-import com.woocommerce.android.ui.orders.creation.customerlistnew.CustomerListGetSupportedSearchModes.Companion.SEARCH_MODE_VALUE_EMAIL
-import com.woocommerce.android.ui.orders.creation.customerlistnew.CustomerListGetSupportedSearchModes.Companion.SEARCH_MODE_VALUE_NAME
-import com.woocommerce.android.ui.orders.creation.customerlistnew.CustomerListGetSupportedSearchModes.Companion.SEARCH_MODE_VALUE_USERNAME
+import com.woocommerce.android.ui.orders.creation.customerlist.CustomerListGetSupportedSearchModes.Companion.SEARCH_MODE_VALUE_ALL
+import com.woocommerce.android.ui.orders.creation.customerlist.CustomerListGetSupportedSearchModes.Companion.SEARCH_MODE_VALUE_EMAIL
+import com.woocommerce.android.ui.orders.creation.customerlist.CustomerListGetSupportedSearchModes.Companion.SEARCH_MODE_VALUE_NAME
+import com.woocommerce.android.ui.orders.creation.customerlist.CustomerListGetSupportedSearchModes.Companion.SEARCH_MODE_VALUE_USERNAME
 import com.woocommerce.android.viewmodel.MultiLiveEvent
 import com.woocommerce.android.viewmodel.ScopedViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
